Do not try to add LUKSDevice in LUKSFormatPopulator
With LUKS 2 we can't just add the LUKSDevice as a child of the
device with LUKSFormat because there can be a dm-integrity
device between these two.
Running format.setup() for unopened devices is enough -- populator
detects the new LUKS device (and the integrity device too) and
will add it later using the LUKSDevicePopulator.